requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
xai-sdk>=0.0.1
rich>=13.7.0
